Rename check_and_prepare_model_process.py
Rename check_and_prepare_model_process.py to check_and_prepare_model_process_structural.py
