+++ applications/StructuralMechanicsApplication/python_scripts/check_and_prepare_model_process_structural.py
